+++ src/main/java/com/example/gistcompetitioncnserver/verification/application/VerificationCodeGenerator.java
@@ -1,4 +1,4 @@
-package com.example.gistcompetitioncnserver.verification;
+package com.example.gistcompetitioncnserver.verification.application;
 
 public interface VerificationCodeGenerator {
     String generate();
