+++ src/main/java/com/example/gistcompetitioncnserver/verification/domain/VerificationInfo.java
@@ -1,4 +1,4 @@
-package com.example.gistcompetitioncnserver.verification;
+package com.example.gistcompetitioncnserver.verification.domain;
 
 import lombok.Getter;
