+++ src/main/java/com/example/gistcompetitioncnserver/verification/dto/VerificationEmailRequest.java
@@ -1,4 +1,4 @@
-package com.example.gistcompetitioncnserver.verification;
+package com.example.gistcompetitioncnserver.verification.dto;
 
 import javax.validation.constraints.Email;
